+++ common/state/src/state.rs
@@ -240,6 +240,7 @@ impl State {
         // Register server -> all clients synced components.
         ecs.register::<comp::Body>();
         ecs.register::<comp::Hardcore>();
+        ecs.register::<comp::CharacterClass>();
         ecs.register::<comp::body::parts::Heads>();
         ecs.register::<comp::Player>();
         ecs.register::<comp::Stats>();